change some methods in services and repositories, develop machine block: machines page, machine_info page, machine_notes page, machine_tasks page, machine_task_info page, machine_alert page, machine_alert_info page, machine_maintenances page, machine_maintenance_info page, machine_weighings page, machine_weighing_info page, machine_downtimes page and machine_regions page

+++ src/main/java/com/agro/agro_project/dao/MachineRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository("machineRepository")
 public interface MachineRepository extends JpaRepository<Machine, Integer> {
-    Optional<Machine> findById(Integer id);
+    Optional<Machine> findById(Short id);
 
     List<Machine> findAll();
 }
